+++ docs/en/docs/release-notes.md
@@ -72,6 +72,7 @@ Where here the post can be directly sent like:
 ### Fixed
 
 - Typing for `settings_module` that was not parsing properly.
+- `requestBody` was not displaying when it was required and enabled.
 
 ## 0.15.2